revise

+++ Competitive_Programming/Computation_Advanced/rapid_count_divisor.hpp
@@ -1,9 +1,9 @@
 //約数の個数をO(n^(1/3))で求めるテク(実装はしてみたが意味あるかな？)
 #include "../header.hpp"
 
-const vector<ll> numset = {2,3,5,7,11,13,17,19,23,29,31,37};
+const vector<long long> numset = {2,3,5,7,11,13,17,19,23,29,31,37};
 
-ll modpow(__int128 x, ll k, ll m){
+long long modpow(__int128 x, long long k, long long m){
     __int128 res = 1;
     while(k){
         if(k & 1){
@@ -16,23 +16,23 @@ ll modpow(__int128 x, ll k, ll m){
 }
 
 // check if n is prime
-bool check(ll n){
+bool check(long long n){
     if(n < 2){
       return false;
     }
-    ll d = n - 1;
-    ll s = 0;
+    long long d = n - 1;
+    long long s = 0;
     while(d % 2 == 0){
         d /= 2;
         s++;
     }
-    for(ll a : numset){
+    for(long long a : numset){
         if(a == n){
           return true;
         }
         if(modpow(a,d,n) != 1){
             bool ok = true;
-            for(ll r=0;r<s;r++){
+            for(long long r=0;r<s;r++){
                 if(modpow(a,d*(1LL << r),n) == n-1){
                     ok = false;
                     break;
@@ -46,9 +46,9 @@ bool check(ll n){
     return true;
 }
 
-ll count_devisor(ll N){
-    ll res = 1;
-	for(ll i=2;i*i*i<=N;i++){
+long long count_devisor(long long N){
+    long long res = 1;
+	for(long long i=2;i*i*i<=N;i++){
 		int cnt = 0;
 		while(N%i == 0){
 			cnt++;
@@ -60,7 +60,7 @@ ll count_devisor(ll N){
         if(check(N)){
             res *= 2;
         }else{
-            ll val = (ll)floor((long double)sqrt(N));
+            long long val = (long long)floor((long double)sqrt(N));
             if(val*val == N){
                 res *= 3;
             }else{
